add category id for tab table

+++ database/migrations/2020_11_25_024651_create_package_tab_table.php
@@ -17,7 +17,7 @@ class CreatePackageTabTable extends Migration
             $table->id();
 
             $table->unsignedBigInteger('package_id')->nullable(false)->index();
-            $table->unsignedInteger('tab_id')->nullable(false)->index();
+            $table->unsignedBigInteger('tab_id')->nullable(false)->index();
         });
     }
